Variable Explorer: Move namespacebrowser to the varexp namespace
