+++ spyderlib/plugins/variableexplorer.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright © 2009-2010 Pierre Raybaut
+# Copyright © 2009- The Spyder Development Team
 # Licensed under the terms of the MIT License
 # (see spyderlib/__init__.py for details)
 
-"""Namespace Browser Plugin"""
+"""Variable Explorer Plugin"""
 
 from spyderlib.qt.QtGui import QGroupBox, QStackedWidget, QVBoxLayout, QWidget
 from spyderlib.qt.QtCore import Signal, Slot
@@ -16,7 +16,7 @@ from spyderlib.config.main import CONF
 from spyderlib.utils import programs
 from spyderlib.plugins import SpyderPluginMixin, PluginConfigPage
 from spyderlib.widgets.externalshell.monitor import REMOTE_SETTINGS
-from spyderlib.widgets.externalshell.namespacebrowser import NamespaceBrowser
+from spyderlib.widgets.varexp.namespacebrowser import NamespaceBrowser
 
 
 class VariableExplorerConfigPage(PluginConfigPage):
