+++ spyderlib/widgets/varexp/namespacebrowser.py
